--- src/com/duke/commands/HelpCommand.java
@@ -1,5 +1,0 @@
-package com.duke.commands;
-
-public class HelpCommand extends Command {
-    public static final String COMMAND_WORD="help";
-}
