--- src/com/duke/commands/IncorrectCommand.java
@@ -1,19 +1,0 @@
-package com.duke.commands;
-
-
-/**
- * Represents an incorrect command. Upon execution, produces some feedback to the user.
- */
-public class IncorrectCommand extends Command {
-
-    public final String feedbackToUser;
-
-    public IncorrectCommand(String feedbackToUser) {
-        this.feedbackToUser = feedbackToUser;
-    }
-
-
-    public void execute() {
-       System.out.println(feedbackToUser);
-    }
-}
