--- src/com/duke/exception/DukeException.java
@@ -1,4 +1,0 @@
-package com.duke.exception;
-
-public class DukeException extends Exception {
-}
