--- src/com/duke/exception/EmptyDescriptionException.java
@@ -1,15 +1,0 @@
-package com.duke.exception;
-
-public class EmptyDescriptionException extends Exception{
-     private String message;
-
-     public EmptyDescriptionException(String message)
-     {
-         this.message=message;
-     }
-
-     @Override
-     public String getMessage(){
-         return message;
-     }
-}
