--- src/com/duke/exception/UndefinedTaskException.java
@@ -1,5 +1,0 @@
-package com.duke.exception;
-
-public class UndefinedTaskException extends Exception{
-
-}
