+++ src/main/java/duke/commands/ExitCommand.java
@@ -1,0 +1,14 @@
+package main.java.duke.commands;
+
+/**
+ * Exit the program.
+ */
+public class ExitCommand extends Command {
+
+    public static final String COMMAND_WORD="bye";
+
+
+    public static boolean isExit(Command command) {
+        return command instanceof ExitCommand; // instanceof returns false if it is null
+    }
+}
