+++ src/main/java/duke/commands/HelpCommand.java
@@ -1,0 +1,5 @@
+package main.java.duke.commands;
+
+public class HelpCommand extends Command {
+    public static final String COMMAND_WORD="help";
+}
