+++ src/main/java/duke/exception/DukeException.java
@@ -1,0 +1,4 @@
+package main.java.duke.exception;
+
+public class DukeException extends Exception {
+}
