+++ src/main/java/duke/exception/EmptyDescriptionException.java
@@ -1,0 +1,15 @@
+package main.java.duke.exception;
+
+public class EmptyDescriptionException extends Exception{
+     private String message;
+
+     public EmptyDescriptionException(String message)
+     {
+         this.message=message;
+     }
+
+     @Override
+     public String getMessage(){
+         return message;
+     }
+}
