+++ src/main/java/duke/exception/IllegalValueException.java
@@ -1,0 +1,13 @@
+package main.java.duke.exception;
+
+/**
+ * Signals that some given data does not fulfill some constraints.
+ */
+public class IllegalValueException extends Exception {
+    /**
+     * @param message should contain relevant information on the failed constraint(s)
+     */
+    public IllegalValueException(String message) {
+        super(message);
+    }
+}
