+++ src/main/java/duke/exception/UndefinedTaskException.java
@@ -1,0 +1,5 @@
+package main.java.duke.exception;
+
+public class UndefinedTaskException extends Exception{
+
+}
